Juego sin validaciones

+++ Category/Category.php
@@ -2,30 +2,32 @@
 
 include_once("../DataBase/DBConnect.php");
 $func = "";
-$func = $_GET['func'];
+if (isset($_GET['func'])) {
+    $func = $_GET['func'];
 
-if ($func) {
-    $categ = new Category();
-    if (strcmp($func, "del") == 0) {//elimnar
-        $id = $_GET['id'];
-        $categ->delete($id);
-    } else if (strcmp($func, "insert") == 0) {//insertar
-        $name = $_POST['name'];
-        $categ->insert($name);
-    } else if (strcmp($func, "edit") == 0) {//editar
-        $id = $_POST['id'];
-        $name = $_POST['name'];
-        $categ->edit($name, $id);
+    if ($func) {
+        $categ = new Category();
+        if (strcmp($func, "del") == 0) {//elimnar
+            $id = $_GET['id'];
+            $categ->delete($id);
+        } else if (strcmp($func, "insert") == 0) {//insertar
+            $name = $_POST['name'];
+            $categ->insert($name);
+        } else if (strcmp($func, "edit") == 0) {//editar
+            $id = $_POST['id'];
+            $name = $_POST['name'];
+            $categ->edit($name, $id);
+        }
+        header('Location:./ListCategory.php '); //devuelve a la pág ListCategory
     }
-    header('Location:./ListCategory.php '); //devuelve a la pág ListCategory
 }
 
 class Category {
 
     public $id;
     public $name;
-    
-     function __construct() {
+
+    function __construct() {
         
     }
 
@@ -34,9 +36,9 @@ class Category {
         $this->name = $name;
     }
 
-    public function select() {
+    public function select($sentenc) {
         $connect = new DBConnect();
-        $result = pg_query($connect->getDB(), "SELECT id,name FROM category");
+        $result = pg_query($connect->getDB(), "SELECT c.id,c.name FROM category c " . (!$sentenc ? "" : "WHERE c.id IN(SELECT s.category_id FROM sentence s)"));
         if (!$result) {
             echo "Ha ocurrido un error.\n";
             exit;
@@ -51,7 +53,9 @@ class Category {
             $categories[$num] = $category;
             $num++;
         }
-        return $categories;
+        if (isset($categories)) {
+            return $categories;
+        }
     }
 
     public function getDataToEdit($idD) {
@@ -87,6 +91,12 @@ class Category {
             echo "Ha ocurrido un error.\n";
             exit;
         }
+    }
+
+    public function getNumSentences($idD) {
+        $connect = new DBConnect();
+        $result = pg_query($connect->getDB(), "SELECT COUNT(*) FROM sentence WHERE category_id=$idD");
+        return pg_fetch_row($result)[0];
     }
 
 }
