+++ GameSentence/GameSentence.php
@@ -2,14 +2,16 @@
 
 include_once("../DataBase/DBConnect.php");
 include_once("../Sentence/Sentence.php");
-$func = "";
-$func = $_GET['func'];
+$func;
 
-if ($func) {
-    $gameSent = new GameSentence();
-    if (strcmp($func, "del") == 0) {//elimnar
-        $id = $_GET['id'];
-        $gameSent->delete($id);
+if (isset($_GET['func'])) {
+    $func = $_GET['func'];
+    if ($func) {
+        $gameSent = new GameSentence();
+        if (strcmp($func, "del") == 0) {//elimnar
+            $id = $_GET['id'];
+            $gameSent->delete($id);
+        }
     }
 }
 
@@ -18,28 +20,18 @@ class GameSentence {
     public $id;
     public $gameId;
     public $sentenceId;
+    public $isRight;
 
     function __construct() {
         
     }
 
-    /* public function getDataToEdit($idD) {
-      $connect = new DBConnect();
-      $result = pg_query($connect->getDB(), "SELECT num_phrases,category_id FROM game WHERE id=" . $idD);
-      $data[0] = pg_fetch_row($result)[0];
-      $data[1] = pg_fetch_row($result)[1];
-      return $data;
-      }
-
-      public function delete($idD) {
-      $connect = new DBConnect();
-      $result = pg_query($connect->getDB(), "DELETE FROM game WHERE id=" . $idD);
-      if (!$result) {
-      echo "Ha ocurrido un error.\n";
-      exit;
-      }
-      return $result;
-      } */
+    public function setPars($id, $gameId, $sentenceId, $isRight) {
+        $this->id = $id;
+        $this->gameId = $gameId;
+        $this->sentenceId = $sentenceId;
+        $this->isRight = $isRight;
+    }
 
     public function insert($gameId, $SentenceId) {
         $connect = new DBConnect();
@@ -51,32 +43,41 @@ class GameSentence {
         return $result;
     }
 
-    public function getNextSentence($gameId) {
+    public function getNextGameSentence($gameId) {
         $connect = new DBConnect();
-        $result = pg_query($connect->getDB(), "SELECT MIN(gs.id)"
+        $result = pg_query($connect->getDB(), "SELECT gss.id,gss.game_id,gss.sentence_id FROM game_sentence gss WHERE gss.id= (SELECT MIN(gs.id) "
                 . "FROM game_sentence gs "
-                . "WHERE gs.game_id=" . $gameId . " AND right IS NULL");
-        $sentenceId = pg_fetch_row($result)[0];
-        $sentence = new Sentence();
-        return $sentence->getDataToEdit($sentenceId);
+                . "WHERE gs.game_id=" . $gameId . " AND gs.is_right IS NULL)");
+        $row = pg_fetch_row($result);
+
+        $gameSent = new GameSentence();
+        $gameSent->setPars($row[0], $row[1], $row[2], null);
+        return $gameSent;
     }
 
     public function setResult($gameId, $right) { //cambia a bien o mal
         $connect = new DBConnect();
-        $maxGameSentence = getMaxGameSentence($gameId);
-        $result = pg_query($connect->getDB(), "UPDATE game_sentence SET right=" . $right . " WHERE id=" . $maxGameSentence);
+        $nextGS = $this->getNextGameSentence($gameId);
+        $result = pg_query($connect->getDB(), "UPDATE game_sentence SET is_right = " . $right . " WHERE id = " . $nextGS->id);
         if (!$result) {
             echo "Ha ocurrido un error.\n";
             exit;
         }
     }
 
-    /*  public function edit($numPhraE, $categId, $idE) {
-      $connect = new DBConnect();
-      $result = pg_query($connect->getDB(), "UPDATE game SET num_phrases=" . $numPhraE . ",category_id=" . $categId . " WHERE id=" . $idE);
-      if (!$result) {
-      echo "Ha ocurrido un error.\n";
-      exit;
-      }
-      } */
+    public function getPunctuation($gameId) {
+        $connect = new DBConnect();
+        $result = pg_query($connect->getDB(), "SELECT COUNT(*),(SELECT COUNT(*) "
+                . "FROM game_sentence gss "
+                . "WHERE gss.game_id=" . $gameId . " "
+                . "AND gss.is_right) "
+                . "FROM game_sentence gs "
+                . "WHERE gs.game_id = " . $gameId);
+        $row = pg_fetch_row($result);
+
+        $punct[0] = $row[0];
+        $punct[1] = $row[1];
+        return $punct;
+    }
+
 }
